refactor(monorepo): move runtime into apps and packages
