+++ .github/workflows/release-binaries.yml
@@ -42,9 +42,9 @@ jobs:
         run: |
           mkdir -p dist/release
 
-          bun build ./bin/cli.ts --compile --target=bun-darwin-x64 --outfile dist/release/bingbong-darwin-x64/bingbong
-          bun build ./bin/cli.ts --compile --target=bun-darwin-arm64 --outfile dist/release/bingbong-darwin-arm64/bingbong
-          bun build ./bin/cli.ts --compile --target=bun-linux-x64 --outfile dist/release/bingbong-linux-x64/bingbong
+          bun build ./packages/cli/bin/cli.ts --compile --target=bun-darwin-x64 --outfile dist/release/bingbong-darwin-x64/bingbong
+          bun build ./packages/cli/bin/cli.ts --compile --target=bun-darwin-arm64 --outfile dist/release/bingbong-darwin-arm64/bingbong
+          bun build ./packages/cli/bin/cli.ts --compile --target=bun-linux-x64 --outfile dist/release/bingbong-linux-x64/bingbong
 
       - name: Package release archives
         run: |
