+++ apps/client/package.json
@@ -7,6 +7,9 @@
     "build": "bunx vite build",
     "typecheck": "bunx tsc --noEmit -p tsconfig.json"
   },
+  "dependencies": {
+    "@bingbong/protocol": "file:../../packages/protocol"
+  },
   "devDependencies": {
     "typescript": "^5.3.0",
     "vite": "^6.0.0"
