+++ packages/cli/bin/cli.ts
